Advent of kotlin. Day 6. Part 2.
- improved solution, instead of calculating each speed and it's distance
  in the range, we're solving standart quadratic equation to find a range
in which covered distances will be greater than given distance.

Here is an explanation:
The equation we're dealing with here is speed * (time - speed). This equation represents a parabola, which is a graph that is shaped like a U or an upside-down U. The parabola you're dealing with in this case is upside-down because of the negative sign between speed and time - speed.

A parabola is symmetrical, meaning it looks the same on both sides. The highest or lowest point of a parabola is called its vertex, and the x-coordinate of this vertex is time / 2 in your case.

The calculation val maxDistance = (maxSpeed * (t - maxSpeed)).toLong() finds the maximum distance you can cover, which is the y-coordinate of the vertex of this parabola. If this maximum distance is greater than your actual distance d, then there must be some lower and higher speeds such that their corresponding distances are exactly d.

The calculation val sqrtPart = Math.sqrt((t * t / 4) - d) is based on solving the equation speed * (time - speed) = d for speed. This equation is a quadratic equation, and solving it involves computing a square root, hence the Math.sqrt part.

Finally, val lowSpeed = Math.ceil(t / 2 - sqrtPart).toLong() and val highSpeed = Math.floor(t / 2 + sqrtPart).toLong() compute the lowest and highest speeds, respectively, such that the covered distance is greater than d. This range of speeds is added to the counter number, because any speed within this range will result in a covered distance that meets your requirement.